+++ cherrypic-api/src/main/java/org/cherrypic/domain/tempalbum/repository/TempAlbumImageRepository.java
@@ -3,5 +3,4 @@ package org.cherrypic.domain.tempalbum.repository;
 import org.cherrypic.tempalbum.entity.TempAlbumImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TempAlbumImageRepository
-        extends JpaRepository<TempAlbumImage, Long>, TempAlbumImageRepositoryCustom {}
+public interface TempAlbumImageRepository extends JpaRepository<TempAlbumImage, Long> {}
